requirements.txt
django
djangorestframework
django-environ
pillow